+++ apps/core-api/src/configs/swagger-option.ts
