+++ apps/core-api/src/configs/swagger-ui-option.ts
