+++ apps/core-api/src/routes/api/v1/interviews/_sessionId/stt-token/index.ts
@@ -6,8 +6,8 @@ import { Static, Type } from '@sinclair/typebox'
 import axios from 'axios'
 import { FastifyInstance, RequestGenericInterface, RouteHandler } from 'fastify'
 
-import { Tag } from '@/configs/swaggerOption'
-import SchemaId from '@/utils/schemaId'
+import { Tag } from '@/configs/swagger-option'
+import SchemaId from '@/utils/schema-id'
 
 const controller: FastifyPluginAsyncTypebox = async (
   fastify: FastifyInstance,
