+++ apps/core-api/src/routes/api/v1/refresh/index.ts
@@ -4,8 +4,8 @@ import {
   RouteShorthandOptions,
 } from 'fastify'
 
-import { Tag } from '@/configs/swaggerOption'
-import SchemaId from '@/utils/schemaId'
+import { Tag } from '@/configs/swagger-option'
+import SchemaId from '@/utils/schema-id'
 
 interface JWTPayload {
   userId: string
