+++ apps/core-api/src/routes/api/v1/users/_userId/index.ts
@@ -6,8 +6,8 @@ import { Prisma } from '@prisma/client'
 import { Static, Type } from '@sinclair/typebox'
 import { FastifyInstance, FastifySchema, RouteHandler } from 'fastify'
 
-import { Tag } from '@/configs/swaggerOption'
-import SchemaId from '@/utils/schemaId'
+import { Tag } from '@/configs/swagger-option'
+import SchemaId from '@/utils/schema-id'
 
 const controller: FastifyPluginAsyncTypebox = async (
   fastify: FastifyInstance,
