+++ apps/core-api/src/utils/schema-id.ts
